app/requirements.txt
uvicorn
filestack-python
fastapi
pandas
numpy
scikit-learn
ipykernel
inflect
nltk
gensim
pandas_profiling
pydantic==1.*
ydata-profiling
nbformat
nbconvert